backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
openai-agents==0.5.0
openai-chatkit==1.1.1
python-dotenv==1.0.1